+++ src/app/instance.ts
@@ -1,10 +1,10 @@
 import { CuiSetup } from "../core/models/setup";
 import { DefaultSetup } from "./defaults/setup";
-import { is, getName, joinAttributesForQuery } from "../core/utlis/functions";
+import { is, getName, joinAttributesForQuery, getSystemLightMode } from "../core/utlis/functions";
 import { ElementManager } from "./managers/element";
-import { CuiLogLevel } from "../core/utlis/types";
+import { CuiLogLevel, CuiClearCacheType, CuiLightMode } from "../core/utlis/types";
 import { CLASSES, MUTATED_ATTRIBUTES, ATTRIBUTES } from "../core/utlis/statics";
-import { IUIInteractionProvider, ICuiDictionary, ICuiLogger } from "../core/models/interfaces";
+import { IUIInteractionProvider, ICuiDictionary, ICuiLogger, ICuiCacheManager } from "../core/models/interfaces";
 import { FastDom } from "../core/utlis/interactions";
 import { CuiInteractionsFactory } from "../core/factories/interactions";
 import { CuiDictionary } from "../core/utlis/dictionary";
@@ -13,7 +13,9 @@ import { CuiAttributeMutationHandler } from "./managers/mutations";
 import { CuiLoggerFactory } from "../core/factories/logger";
 import { CuiInstanceColorHandler } from "./handlers/colors";
 import { CuiToastHandler } from "./managers/toast";
-import { ListManager } from "./managers/list";
+import { CollectionManager } from "./managers/collection";
+import { CuiCacheManager } from "./managers/cache";
+import { LightModeListener } from "../core/listeners/light";
 
 export class CuiInstance {
     #log: ICuiLogger;
@@ -21,9 +23,12 @@ export class CuiInstance {
     #logLevel: CuiLogLevel;
     #interactions: IUIInteractionProvider;
     #cache: ICuiDictionary<ElementManager>;
-
+    #collectionCache: ICuiDictionary<CollectionManager>;
     #mutationObserver: ICuiMutionObserver;
     #toastManager: CuiToastHandler;
+    #cacheManager: ICuiCacheManager;
+    #lightModeListener: LightModeListener;
+    autoLightMode: boolean;
 
     //pubic
     colors: CuiInstanceColorHandler;
@@ -33,9 +38,13 @@ export class CuiInstance {
         this.#interactions = CuiInteractionsFactory.get(setup.interaction)
         this.#log = CuiLoggerFactory.get('CuiInstance', this.#logLevel)
         this.#cache = new CuiDictionary<ElementManager>();
+        this.#collectionCache = new CuiDictionary<CollectionManager>();
+        this.#cacheManager = new CuiCacheManager(setup.cacheSize ?? DefaultSetup.cacheSize);
         this.#mutationObserver = new CuiMutationObserver(document.body, this.#interactions)
         this.colors = new CuiInstanceColorHandler(this.#interactions)
         this.#toastManager = new CuiToastHandler(this.#interactions, this.#prefix, setup.animationTimeLong ?? DefaultSetup.animationTimeLong);
+        this.#lightModeListener = new LightModeListener(this);
+        this.autoLightMode = setup.autoLightMode ?? DefaultSetup.autoLightMode;
     }
 
     init(icons?: any): CuiInstance {
@@ -58,6 +67,10 @@ export class CuiInstance {
             attributeFilter: MUTATED_ATTRIBUTES
         });
         this.#mutationObserver.start();
+        this.#lightModeListener.start();
+        if (this.autoLightMode && getSystemLightMode() === 'dark') {
+            this.setLightMode('dark')
+        }
         return this;
     }
 
@@ -67,37 +80,50 @@ export class CuiInstance {
 
 
     get(selector: string): ElementManager {
-        const existing = this.#cache.get(selector);
+        const existing = this.#cacheManager.get(selector);
         if (is(existing)) {
-            if (existing.refresh()) {
-                return existing;
-            }
-            this.#cache.remove(selector);
+            return existing as ElementManager;
         }
         const elements = this.all(selector);
         if (!elements) {
             return undefined
         }
         const newElement = new ElementManager(elements, this.#interactions, this.#logLevel);
-        this.#cache.add(selector, newElement)
+        this.#cacheManager.put(selector, newElement)
         return newElement
     }
 
-    list(selector: string): ListManager {
+    collection(selector: string): CollectionManager {
+        const existing = this.#cacheManager.get(selector);
+        if (is(existing)) {
+            return existing as CollectionManager;
+        }
         const elements = this.all(selector);
         if (!is(elements)) {
             return undefined;
         }
-        return new ListManager(elements, this.#interactions, this.#logLevel);
+        let manager = new CollectionManager(elements, this.#interactions, this.#logLevel);
+        this.#cacheManager.put(selector, manager)
+        return manager;
     }
 
-    toggleDarkMode(): void {
-        const name: string = getName(this.#prefix, CLASSES.dark)
-        const classes = document.body.classList
-        if (classes.contains(name)) {
-            classes.remove(name)
-        } else {
-            classes.add(name)
+    // toggleDarkMode(): void {
+    //     const name: string = getName(this.#prefix, CLASSES.dark)
+    //     const classes = document.body.classList
+    //     if (classes.contains(name)) {
+    //         classes.remove(name)
+    //     } else {
+    //         classes.add(name)
+    //     }
+    // }
+
+    setLightMode(mode: CuiLightMode) {
+        const name: string = getName(this.#prefix, CLASSES.dark);
+        const classes = document.body.classList;
+        if (mode === 'dark' && !classes.contains(name)) {
+            classes.add(name);
+        } else if (mode === 'light' && classes.contains(name)) {
+            classes.remove(name);
         }
     }
 
@@ -118,5 +144,19 @@ export class CuiInstance {
             return undefined;
         }
         return [...nodes];
+    }
+
+    clearCache(clearType: CuiClearCacheType): void {
+        switch (clearType) {
+            case 'element':
+                this.#cache.clear();
+                break;
+            case 'collection':
+                this.#collectionCache.clear();
+                break;
+            case 'all':
+                this.#cache.clear();
+                this.#collectionCache.clear();
+        }
     }
 }