+++ src/core/utlis/types.ts
@@ -1,3 +1,5 @@
 export type CuiLogLevel = "none" | "error" | "warning" | "debug";
 export type CuiInteractionsType = 'sync' | 'async';
-export type CuiColorSetType = 'light' | 'dark' | 'accent' | 'secondary' | 'success' | 'warning' | 'error'
+export type CuiColorSetType = 'light' | 'dark' | 'accent' | 'secondary' | 'success' | 'warning' | 'error';
+export type CuiClearCacheType = 'element' | "collection" | "all";
+export type CuiLightMode = 'light' | 'dark'